+++ src/types/PaymentMethod.ts
@@ -283,6 +283,7 @@ export type Type =
   | 'Oxxo'
   | 'USBankAccount'
   | 'PayPal'
+  | 'PayByBank'
   | 'Unknown';
 
 export type CollectBankAccountParams = {